semana1_python/pyproject.toml
[project]
name = "grificos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pysdl2>=0.9.17",
    "pysdl2-dll>=2.32.0",
]
